Separate project lib files
