--- lib/projects.ts
@@ -1,1165 +1,0 @@
-import fs from 'fs-extra';
-import path from 'path';
-import archiver from 'archiver';
-import tmp, { FileResult } from 'tmp';
-import chalk from 'chalk';
-import findup from 'findup-sync';
-import { logger } from '@hubspot/local-dev-lib/logger';
-import { getEnv } from '@hubspot/local-dev-lib/config';
-import { getHubSpotWebsiteOrigin } from '@hubspot/local-dev-lib/urls';
-import { fetchFileFromRepository } from '@hubspot/local-dev-lib/github';
-import { ENVIRONMENTS } from '@hubspot/local-dev-lib/constants/environments';
-import {
-  createProject,
-  getBuildStatus,
-  getBuildStructure,
-  getDeployStatus,
-  getDeployStructure,
-  fetchProject,
-  uploadProject,
-  fetchBuildWarnLogs,
-  fetchDeployWarnLogs,
-} from '@hubspot/local-dev-lib/api/projects';
-import { isSpecifiedError } from '@hubspot/local-dev-lib/errors/index';
-import { shouldIgnoreFile } from '@hubspot/local-dev-lib/ignoreRules';
-import { getCwd, getAbsoluteFilePath } from '@hubspot/local-dev-lib/path';
-import { downloadGithubRepoContents } from '@hubspot/local-dev-lib/github';
-import { RepoPath } from '@hubspot/local-dev-lib/types/Github';
-import {
-  Project,
-  WarnLogsResponse,
-} from '@hubspot/local-dev-lib/types/Project';
-import { HubSpotPromise } from '@hubspot/local-dev-lib/types/Http';
-import { ComponentStructureResponse } from '@hubspot/local-dev-lib/types/ComponentStructure';
-import { Build } from '@hubspot/local-dev-lib/types/Build';
-import { Deploy } from '@hubspot/local-dev-lib/types/Deploy';
-
-import {
-  FEEDBACK_INTERVAL,
-  POLLING_DELAY,
-  PROJECT_BUILD_TEXT,
-  PROJECT_DEPLOY_TEXT,
-  PROJECT_CONFIG_FILE,
-  PROJECT_TASK_TYPES,
-  PROJECT_ERROR_TYPES,
-  HUBSPOT_PROJECT_COMPONENTS_GITHUB_PATH,
-  PROJECT_COMPONENT_TYPES,
-} from './constants';
-import { promptUser } from './prompts/promptUtils';
-import { EXIT_CODES } from './enums/exitCodes';
-import { uiLine, uiLink, uiAccountDescription, uiCommandReference } from './ui';
-import { i18n } from './lang';
-import SpinniesManager from './ui/SpinniesManager';
-import { logError, ApiErrorContext } from './errorHandlers/index';
-import {
-  ProjectTemplate,
-  ProjectConfig,
-  ProjectPollStatusFunctionText,
-  ProjectTask,
-  ProjectSubtask,
-  ProjectAddComponentData,
-  ProjectTemplateRepoConfig,
-  ComponentTemplate,
-} from '../types/projects';
-
-const i18nKey = 'lib.projects';
-
-const SPINNER_STATUS = {
-  SPINNING: 'spinning',
-};
-
-function writeProjectConfig(
-  configPath: string,
-  config: ProjectConfig
-): boolean {
-  try {
-    fs.ensureFileSync(configPath);
-    fs.writeFileSync(configPath, JSON.stringify(config, null, 2));
-    logger.debug(`Wrote project config at ${configPath}`);
-  } catch (e) {
-    logger.debug(e);
-    return false;
-  }
-  return true;
-}
-
-function getIsInProject(dir?: string): boolean {
-  const configPath = getProjectConfigPath(dir);
-  return !!configPath;
-}
-
-function getProjectConfigPath(dir?: string): string | null {
-  const projectDir = dir ? getAbsoluteFilePath(dir) : getCwd();
-
-  const configPath = findup(PROJECT_CONFIG_FILE, {
-    cwd: projectDir,
-    nocase: true,
-  });
-
-  return configPath;
-}
-
-export async function getProjectConfig(dir?: string): Promise<{
-  projectDir: string | null;
-  projectConfig: ProjectConfig | null;
-}> {
-  const configPath = await getProjectConfigPath(dir);
-  if (!configPath) {
-    return { projectConfig: null, projectDir: null };
-  }
-
-  try {
-    const config = fs.readFileSync(configPath);
-    const projectConfig: ProjectConfig = JSON.parse(config.toString());
-    return {
-      projectDir: path.dirname(configPath),
-      projectConfig,
-    };
-  } catch (e) {
-    logger.error('Could not read from project config');
-    return { projectConfig: null, projectDir: null };
-  }
-}
-
-async function createProjectConfig(
-  projectPath: string,
-  projectName: string,
-  template: ProjectTemplate,
-  templateSource: RepoPath,
-  githubRef: string
-): Promise<boolean> {
-  const { projectConfig, projectDir } = await getProjectConfig(projectPath);
-
-  if (projectConfig) {
-    logger.warn(
-      projectPath === projectDir
-        ? 'A project already exists in that location.'
-        : `Found an existing project definition in ${projectDir}.`
-    );
-
-    const { shouldContinue } = await promptUser<{ shouldContinue: boolean }>([
-      {
-        name: 'shouldContinue',
-        message: () => {
-          return projectPath === projectDir
-            ? 'Do you want to overwrite the existing project definition with a new one?'
-            : `Continue creating a new project in ${projectPath}?`;
-        },
-        type: 'confirm',
-        default: false,
-      },
-    ]);
-
-    if (!shouldContinue) {
-      return false;
-    }
-  }
-
-  const projectConfigPath = path.join(projectPath, PROJECT_CONFIG_FILE);
-
-  logger.log(
-    `Creating project config in ${
-      projectPath ? projectPath : 'the current folder'
-    }`
-  );
-
-  const hasCustomTemplateSource = Boolean(templateSource);
-
-  await downloadGithubRepoContents(
-    templateSource || HUBSPOT_PROJECT_COMPONENTS_GITHUB_PATH,
-    template.path,
-    projectPath,
-    hasCustomTemplateSource ? undefined : githubRef
-  );
-  const _config: ProjectConfig = JSON.parse(
-    fs.readFileSync(projectConfigPath).toString()
-  );
-  writeProjectConfig(projectConfigPath, {
-    ..._config,
-    name: projectName,
-  });
-
-  if (template.name === 'no-template') {
-    fs.ensureDirSync(path.join(projectPath, 'src'));
-  }
-
-  return true;
-}
-
-function validateProjectConfig(
-  projectConfig: ProjectConfig,
-  projectDir: string
-): void {
-  if (!projectConfig) {
-    logger.error(
-      i18n(`${i18nKey}.validateProjectConfig.configNotFound`, {
-        createCommand: uiCommandReference('hs project create'),
-      })
-    );
-    return process.exit(EXIT_CODES.ERROR);
-  }
-
-  if (!projectConfig.name || !projectConfig.srcDir) {
-    logger.error(i18n(`${i18nKey}.validateProjectConfig.configMissingFields`));
-    return process.exit(EXIT_CODES.ERROR);
-  }
-
-  const resolvedPath = path.resolve(projectDir, projectConfig.srcDir);
-  if (!resolvedPath.startsWith(projectDir)) {
-    const projectConfigFile = path.relative(
-      '.',
-      path.join(projectDir, PROJECT_CONFIG_FILE)
-    );
-    logger.error(
-      i18n(`${i18nKey}.validateProjectConfig.srcOutsideProjectDir`, {
-        srcDir: projectConfig.srcDir,
-        projectConfig: projectConfigFile,
-      })
-    );
-    return process.exit(EXIT_CODES.ERROR);
-  }
-
-  if (!fs.existsSync(resolvedPath)) {
-    logger.error(
-      i18n(`${i18nKey}.validateProjectConfig.srcDirNotFound`, {
-        srcDir: projectConfig.srcDir,
-        projectDir: projectDir,
-      })
-    );
-
-    return process.exit(EXIT_CODES.ERROR);
-  }
-}
-
-async function pollFetchProject(
-  accountId: number,
-  projectName: string
-): HubSpotPromise<Project> {
-  // Temporary solution for gating slowness. Retry on 403 statusCode
-  return new Promise((resolve, reject) => {
-    let pollCount = 0;
-    SpinniesManager.init();
-    SpinniesManager.add('pollFetchProject', {
-      text: i18n(`${i18nKey}.pollFetchProject.checkingProject`, {
-        accountIdentifier: uiAccountDescription(accountId),
-      }),
-    });
-    const pollInterval = setInterval(async () => {
-      try {
-        const response = await fetchProject(accountId, projectName);
-        if (response && response.data) {
-          SpinniesManager.remove('pollFetchProject');
-          clearInterval(pollInterval);
-          resolve(response);
-        }
-      } catch (err) {
-        if (
-          isSpecifiedError(err, {
-            statusCode: 403,
-            category: 'GATED',
-            subCategory: 'BuildPipelineErrorType.PORTAL_GATED',
-          }) &&
-          pollCount < 15
-        ) {
-          pollCount += 1;
-        } else {
-          SpinniesManager.remove('pollFetchProject');
-          clearInterval(pollInterval);
-          reject(err);
-        }
-      }
-    }, POLLING_DELAY);
-  });
-}
-
-async function ensureProjectExists(
-  accountId: number,
-  projectName: string,
-  {
-    forceCreate = false,
-    allowCreate = true,
-    noLogs = false,
-    withPolling = false,
-    uploadCommand = false,
-  } = {}
-): Promise<{
-  projectExists: boolean;
-  project?: Project;
-}> {
-  const accountIdentifier = uiAccountDescription(accountId);
-  try {
-    const { data: project } = withPolling
-      ? await pollFetchProject(accountId, projectName)
-      : await fetchProject(accountId, projectName);
-    return { projectExists: !!project, project };
-  } catch (err) {
-    if (isSpecifiedError(err, { statusCode: 404 })) {
-      let shouldCreateProject = forceCreate;
-      if (allowCreate && !shouldCreateProject) {
-        const promptKey = uploadCommand ? 'createPromptUpload' : 'createPrompt';
-        const promptResult = await promptUser<{ shouldCreateProject: boolean }>(
-          [
-            {
-              name: 'shouldCreateProject',
-              message: i18n(`${i18nKey}.ensureProjectExists.${promptKey}`, {
-                projectName,
-                accountIdentifier,
-              }),
-              type: 'confirm',
-            },
-          ]
-        );
-        shouldCreateProject = promptResult.shouldCreateProject;
-      }
-
-      if (shouldCreateProject) {
-        try {
-          const { data: project } = await createProject(accountId, projectName);
-          logger.success(
-            i18n(`${i18nKey}.ensureProjectExists.createSuccess`, {
-              projectName,
-              accountIdentifier,
-            })
-          );
-          return { projectExists: true, project };
-        } catch (err) {
-          logError(err, new ApiErrorContext({ accountId }));
-          return { projectExists: false };
-        }
-      } else {
-        if (!noLogs) {
-          logger.log(
-            i18n(`${i18nKey}.ensureProjectExists.notFound`, {
-              projectName,
-              accountIdentifier,
-            })
-          );
-        }
-        return { projectExists: false };
-      }
-    }
-    if (
-      isSpecifiedError(err, {
-        statusCode: 401,
-      })
-    ) {
-      logger.error(err.message);
-      process.exit(EXIT_CODES.ERROR);
-    }
-    logError(err, new ApiErrorContext({ accountId }));
-    process.exit(EXIT_CODES.ERROR);
-  }
-}
-
-function getProjectHomeUrl(accountId: number): string {
-  const baseUrl = getHubSpotWebsiteOrigin(
-    getEnv(accountId) === 'qa' ? ENVIRONMENTS.QA : ENVIRONMENTS.PROD
-  );
-
-  return `${baseUrl}/developer-projects/${accountId}`;
-}
-
-function getProjectDetailUrl(
-  projectName: string,
-  accountId: number
-): string | undefined {
-  if (!projectName) return;
-  return `${getProjectHomeUrl(accountId)}/project/${projectName}`;
-}
-
-function getProjectActivityUrl(projectName: string, accountId: number): string {
-  return `${getProjectDetailUrl(projectName, accountId)}/activity`;
-}
-
-function getProjectBuildDetailUrl(
-  projectName: string,
-  buildId: number,
-  accountId: number
-): string {
-  return `${getProjectActivityUrl(projectName, accountId)}/build/${buildId}`;
-}
-
-function getProjectDeployDetailUrl(
-  projectName: string,
-  deployId: number,
-  accountId: number
-): string {
-  return `${getProjectActivityUrl(projectName, accountId)}/deploy/${deployId}`;
-}
-
-async function uploadProjectFiles(
-  accountId: number,
-  projectName: string,
-  filePath: string,
-  uploadMessage: string,
-  platformVersion: string
-): Promise<{ buildId?: number; error: unknown }> {
-  SpinniesManager.init({});
-  const accountIdentifier = uiAccountDescription(accountId);
-
-  SpinniesManager.add('upload', {
-    text: i18n(`${i18nKey}.uploadProjectFiles.add`, {
-      accountIdentifier,
-      projectName,
-    }),
-    succeedColor: 'white',
-  });
-
-  let buildId: number | undefined;
-  let error: unknown;
-
-  try {
-    const { data: upload } = await uploadProject(
-      accountId,
-      projectName,
-      filePath,
-      uploadMessage,
-      platformVersion
-    );
-
-    buildId = upload.buildId;
-
-    SpinniesManager.succeed('upload', {
-      text: i18n(`${i18nKey}.uploadProjectFiles.succeed`, {
-        accountIdentifier,
-        projectName,
-      }),
-    });
-
-    logger.debug(
-      i18n(`${i18nKey}.uploadProjectFiles.buildCreated`, {
-        buildId,
-        projectName,
-      })
-    );
-  } catch (err) {
-    SpinniesManager.fail('upload', {
-      text: i18n(`${i18nKey}.uploadProjectFiles.fail`, {
-        accountIdentifier,
-        projectName,
-      }),
-    });
-
-    error = err;
-  }
-
-  return { buildId, error };
-}
-
-type ProjectPollResult = {
-  succeeded: boolean;
-  buildId: number;
-  buildResult: Build;
-  deployResult: Deploy | null;
-};
-
-async function pollProjectBuildAndDeploy(
-  accountId: number,
-  projectConfig: ProjectConfig,
-  tempFile: FileResult,
-  buildId: number,
-  silenceLogs = false
-): Promise<ProjectPollResult> {
-  let buildStatus = await pollBuildStatus(
-    accountId,
-    projectConfig.name,
-    buildId,
-    null,
-    silenceLogs
-  );
-
-  if (!silenceLogs) {
-    uiLine();
-  }
-
-  const result: ProjectPollResult = {
-    succeeded: true,
-    buildId,
-    buildResult: buildStatus,
-    deployResult: null,
-  };
-
-  if (buildStatus.status === 'FAILURE') {
-    result.succeeded = false;
-    return result;
-  } else if (buildStatus.isAutoDeployEnabled) {
-    if (!silenceLogs) {
-      logger.log(
-        i18n(
-          `${i18nKey}.pollProjectBuildAndDeploy.buildSucceededAutomaticallyDeploying`,
-          {
-            accountIdentifier: uiAccountDescription(accountId),
-            buildId,
-          }
-        )
-      );
-
-      await displayWarnLogs(accountId, projectConfig.name, buildId);
-    }
-
-    // autoDeployId of 0 indicates a skipped deploy
-    const getIsDeploying = () =>
-      buildStatus.autoDeployId > 0 && buildStatus.deployStatusTaskLocator;
-
-    // Sometimes the deploys do not immediately initiate, give them a chance to kick off
-    if (!getIsDeploying()) {
-      buildStatus = await pollBuildAutodeployStatus(
-        accountId,
-        projectConfig.name,
-        buildId
-      );
-    }
-
-    if (getIsDeploying()) {
-      const deployStatus = await pollDeployStatus(
-        accountId,
-        projectConfig.name,
-        Number(buildStatus.deployStatusTaskLocator.id),
-        buildId,
-        silenceLogs
-      );
-      result.deployResult = deployStatus;
-
-      if (deployStatus.status === 'FAILURE') {
-        result.succeeded = false;
-      }
-    } else if (!silenceLogs) {
-      logger.log(
-        i18n(
-          `${i18nKey}.pollProjectBuildAndDeploy.unableToFindAutodeployStatus`,
-          {
-            buildId,
-            viewDeploysLink: uiLink(
-              i18n(`${i18nKey}.pollProjectBuildAndDeploy.viewDeploys`),
-              getProjectActivityUrl(projectConfig.name, accountId)
-            ),
-          }
-        )
-      );
-    }
-  }
-
-  try {
-    if (tempFile) {
-      tempFile.removeCallback();
-      logger.debug(
-        i18n(`${i18nKey}.pollProjectBuildAndDeploy.cleanedUpTempFile`, {
-          path: tempFile.name,
-        })
-      );
-    }
-  } catch (e) {
-    logger.error(e);
-  }
-
-  if (result && result.deployResult) {
-    await displayWarnLogs(
-      accountId,
-      projectConfig.name,
-      result.deployResult.deployId,
-      true
-    );
-  }
-  return result;
-}
-
-type ProjectUploadCallbackFunction<T> = (
-  accountId: number,
-  projectConfig: ProjectConfig,
-  tempFile: FileResult,
-  buildId?: number
-) => Promise<T | undefined>;
-
-type ProjectUploadDefaultResult = {
-  uploadError?: unknown;
-};
-
-async function handleProjectUpload<T = ProjectUploadDefaultResult>(
-  accountId: number,
-  projectConfig: ProjectConfig,
-  projectDir: string,
-  callbackFunc: ProjectUploadCallbackFunction<T>,
-  uploadMessage: string
-) {
-  const srcDir = path.resolve(projectDir, projectConfig.srcDir);
-
-  const filenames = fs.readdirSync(srcDir);
-  if (!filenames || filenames.length === 0) {
-    logger.log(
-      i18n(`${i18nKey}.handleProjectUpload.emptySource`, {
-        srcDir: projectConfig.srcDir,
-      })
-    );
-    process.exit(EXIT_CODES.SUCCESS);
-  }
-
-  const tempFile = tmp.fileSync({ postfix: '.zip' });
-
-  logger.debug(
-    i18n(`${i18nKey}.handleProjectUpload.compressing`, {
-      path: tempFile.name,
-    })
-  );
-
-  const output = fs.createWriteStream(tempFile.name);
-  const archive = archiver('zip');
-
-  const result = new Promise(resolve =>
-    output.on('close', async function () {
-      let uploadResult: ProjectUploadDefaultResult | T | undefined;
-
-      logger.debug(
-        i18n(`${i18nKey}.handleProjectUpload.compressed`, {
-          byteCount: archive.pointer(),
-        })
-      );
-
-      const { buildId, error } = await uploadProjectFiles(
-        accountId,
-        projectConfig.name,
-        tempFile.name,
-        uploadMessage,
-        projectConfig.platformVersion
-      );
-
-      if (error) {
-        uploadResult = { uploadError: error };
-      } else if (callbackFunc) {
-        uploadResult = await callbackFunc(
-          accountId,
-          projectConfig,
-          tempFile,
-          buildId
-        );
-      }
-      resolve(uploadResult || {});
-    })
-  );
-
-  archive.pipe(output);
-
-  let loggedIgnoredNodeModule = false;
-
-  archive.directory(srcDir, false, file => {
-    const ignored = shouldIgnoreFile(file.name, true);
-    if (ignored) {
-      const isNodeModule = file.name.includes('node_modules');
-
-      if (!isNodeModule || !loggedIgnoredNodeModule) {
-        logger.debug(
-          i18n(`${i18nKey}.handleProjectUpload.fileFiltered`, {
-            filename: file.name,
-          })
-        );
-      }
-
-      if (isNodeModule && !loggedIgnoredNodeModule) {
-        loggedIgnoredNodeModule = true;
-      }
-    }
-    return ignored ? false : file;
-  });
-
-  archive.finalize();
-
-  return result;
-}
-
-function getSubtasks(task: ProjectTask): ProjectSubtask[] {
-  if ('subbuildStatuses' in task) {
-    return task.subbuildStatuses;
-  }
-  return task.subdeployStatuses;
-}
-
-function getSubtaskName(task: ProjectSubtask): string {
-  if ('buildName' in task) {
-    return task.buildName;
-  }
-  return task.deployName;
-}
-
-function getSubtaskType(task: ProjectSubtask): string {
-  if ('buildType' in task) {
-    return task.buildType;
-  }
-  return task.deployType;
-}
-
-type PollTaskStatusFunctionConfig<T extends ProjectTask> = {
-  statusFn: (
-    accountId: number,
-    projectName: string,
-    taskId: number
-  ) => HubSpotPromise<T>;
-  structureFn: (
-    accountId: number,
-    projectName: string,
-    taskId: number
-  ) => HubSpotPromise<ComponentStructureResponse>;
-  statusText: ProjectPollStatusFunctionText;
-  statusStrings: PollTaskStatusStrings;
-  linkToHubSpot: (
-    accountId: number,
-    taskName: string,
-    taskId: number,
-    deployedBuildId: number
-  ) => void;
-};
-
-type PollTaskStatus = 'INITIALIZE' | 'SUCCESS' | 'FAIL' | 'SUBTASK_FAIL';
-
-type PollTaskStatusStringFunction = (name: string, taskId: number) => string;
-
-type PollTaskStatusStrings = {
-  [k in PollTaskStatus]: PollTaskStatusStringFunction;
-};
-
-type PollTaskStatusFunction<T extends ProjectTask> = (
-  accountId: number,
-  taskName: string,
-  taskId: number,
-  deployedBuildId: number | null,
-  silenceLogs: boolean
-) => Promise<T>;
-
-function makePollTaskStatusFunc<T extends ProjectTask>({
-  statusFn,
-  structureFn,
-  statusText,
-  statusStrings,
-  linkToHubSpot,
-}: PollTaskStatusFunctionConfig<T>): PollTaskStatusFunction<T> {
-  return async function (
-    accountId,
-    taskName,
-    taskId,
-    deployedBuildId = null,
-    silenceLogs = false
-  ) {
-    const displayId = deployedBuildId || taskId;
-
-    if (linkToHubSpot && !silenceLogs && deployedBuildId) {
-      logger.log(
-        `\n${linkToHubSpot(accountId, taskName, taskId, deployedBuildId)}\n`
-      );
-    }
-
-    SpinniesManager.init();
-
-    const overallTaskSpinniesKey = `overallTaskStatus-${statusText.STATUS_TEXT}`;
-
-    SpinniesManager.add(overallTaskSpinniesKey, {
-      text: 'Beginning',
-      succeedColor: 'white',
-      failColor: 'white',
-      failPrefix: chalk.bold('!'),
-    });
-
-    const [
-      { data: initialTaskStatus },
-      {
-        data: { topLevelComponentsWithChildren: taskStructure },
-      },
-    ] = await Promise.all([
-      statusFn(accountId, taskName, taskId),
-      structureFn(accountId, taskName, taskId),
-    ]);
-
-    const subtasks = getSubtasks(initialTaskStatus);
-
-    const tasksById = subtasks.reduce(
-      (acc: { [key: string]: ProjectSubtask }, subtask) => {
-        const { id, visible } = subtask;
-        if (visible) {
-          acc[id] = subtask;
-        }
-        return acc;
-      },
-      {}
-    );
-
-    const structuredTasks = Object.keys(taskStructure).map(key => {
-      return {
-        ...tasksById[key],
-        subtasks: taskStructure[key]
-          .filter(taskId => Boolean(tasksById[taskId]))
-          .map(taskId => tasksById[taskId]),
-      };
-    });
-
-    const numComponents = structuredTasks.length;
-    const componentCountText = silenceLogs
-      ? ''
-      : i18n(
-          numComponents === 1
-            ? `${i18nKey}.makePollTaskStatusFunc.componentCountSingular`
-            : `${i18nKey}.makePollTaskStatusFunc.componentCount`,
-          { numComponents }
-        ) + '\n';
-
-    SpinniesManager.update(overallTaskSpinniesKey, {
-      text: `${statusStrings.INITIALIZE(
-        taskName,
-        displayId
-      )}\n${componentCountText}`,
-    });
-
-    if (!silenceLogs) {
-      function addTaskSpinner(
-        subtask: ProjectSubtask,
-        indent: number,
-        newline: boolean
-      ): void {
-        const taskName = getSubtaskName(subtask);
-        const taskType = getSubtaskType(subtask);
-        const formattedTaskType = PROJECT_TASK_TYPES[taskType]
-          ? `[${PROJECT_TASK_TYPES[taskType]}]`
-          : '';
-        const text = `${indent <= 2 ? statusText.STATUS_TEXT : ''} ${chalk.bold(
-          taskName
-        )} ${formattedTaskType} ...${newline ? '\n' : ''}`;
-
-        SpinniesManager.add(subtask.id, {
-          text,
-          indent,
-          succeedColor: 'white',
-          failColor: 'white',
-        });
-      }
-
-      structuredTasks.forEach(task => {
-        addTaskSpinner(task, 2, !task.subtasks || task.subtasks.length === 0);
-        task.subtasks.forEach((subtask, i) =>
-          addTaskSpinner(subtask, 4, i === task.subtasks.length - 1)
-        );
-      });
-    }
-
-    return new Promise<T>((resolve, reject) => {
-      const pollInterval = setInterval(async () => {
-        let taskStatus: T;
-        try {
-          const { data } = await statusFn(accountId, taskName, taskId);
-          taskStatus = data;
-        } catch (e) {
-          logger.debug(e);
-          logError(
-            e,
-            new ApiErrorContext({
-              accountId,
-              projectName: taskName,
-            })
-          );
-          return reject(
-            new Error(
-              i18n(
-                `${i18nKey}.makePollTaskStatusFunc.errorFetchingTaskStatus`,
-                {
-                  taskType:
-                    statusText.TYPE_KEY === PROJECT_BUILD_TEXT.TYPE_KEY
-                      ? 'build'
-                      : 'deploy',
-                }
-              )
-            )
-          );
-        }
-
-        const subtasks = getSubtasks(taskStatus);
-
-        if (!taskStatus || !taskStatus.status || !subtasks) {
-          return reject(
-            new Error(
-              i18n(
-                `${i18nKey}.makePollTaskStatusFunc.errorFetchingTaskStatus`,
-                {
-                  taskType:
-                    statusText.TYPE_KEY === PROJECT_BUILD_TEXT.TYPE_KEY
-                      ? 'build'
-                      : 'deploy',
-                }
-              )
-            )
-          );
-        }
-
-        const { status } = taskStatus;
-
-        if (SpinniesManager.hasActiveSpinners()) {
-          subtasks.forEach(subtask => {
-            const { id, status } = subtask;
-            const spinner = SpinniesManager.pick(id);
-
-            if (!spinner || spinner.status !== SPINNER_STATUS.SPINNING) {
-              return;
-            }
-
-            const topLevelTask = structuredTasks.find(t => t.id == id);
-
-            if (
-              status === statusText.STATES.SUCCESS ||
-              status === statusText.STATES.FAILURE
-            ) {
-              const taskStatusText =
-                subtask.status === statusText.STATES.SUCCESS
-                  ? i18n(`${i18nKey}.makePollTaskStatusFunc.successStatusText`)
-                  : i18n(`${i18nKey}.makePollTaskStatusFunc.failedStatusText`);
-              const hasNewline =
-                spinner?.text?.includes('\n') || Boolean(topLevelTask);
-              const updatedText = `${spinner?.text?.replace(
-                '\n',
-                ''
-              )} ${taskStatusText}${hasNewline ? '\n' : ''}`;
-
-              if (status === statusText.STATES.SUCCESS) {
-                SpinniesManager.succeed(id, { text: updatedText });
-              } else {
-                SpinniesManager.fail(id, { text: updatedText });
-              }
-
-              if (topLevelTask) {
-                topLevelTask.subtasks.forEach(currentSubtask =>
-                  SpinniesManager.remove(currentSubtask.id)
-                );
-              }
-            }
-          });
-
-          if (status === statusText.STATES.SUCCESS) {
-            SpinniesManager.succeed(overallTaskSpinniesKey, {
-              text: statusStrings.SUCCESS(taskName, displayId),
-            });
-            clearInterval(pollInterval);
-            resolve(taskStatus);
-          } else if (status === statusText.STATES.FAILURE) {
-            SpinniesManager.fail(overallTaskSpinniesKey, {
-              text: statusStrings.FAIL(taskName, displayId),
-            });
-
-            if (!silenceLogs) {
-              const failedSubtasks = subtasks.filter(
-                subtask => subtask.status === 'FAILURE'
-              );
-
-              uiLine();
-              logger.log(
-                `${statusStrings.SUBTASK_FAIL(
-                  failedSubtasks.length === 1
-                    ? getSubtaskName(failedSubtasks[0])
-                    : failedSubtasks.length + ' components',
-                  displayId
-                )}\n`
-              );
-              logger.log('See below for a summary of errors.');
-              uiLine();
-
-              const displayErrors = failedSubtasks.filter(
-                subtask =>
-                  subtask?.standardError?.subCategory !==
-                    PROJECT_ERROR_TYPES.SUBBUILD_FAILED &&
-                  subtask?.standardError?.subCategory !==
-                    PROJECT_ERROR_TYPES.SUBDEPLOY_FAILED
-              );
-
-              displayErrors.forEach(subTask => {
-                logger.log(
-                  `\n--- ${chalk.bold(
-                    getSubtaskName(subTask)
-                  )} failed with the following error ---`
-                );
-                logger.error(subTask.errorMessage);
-
-                // Log nested errors
-                if (subTask.standardError && subTask.standardError.errors) {
-                  logger.log();
-                  subTask.standardError.errors.forEach(error => {
-                    logger.log(error.message);
-                  });
-                }
-              });
-            }
-            clearInterval(pollInterval);
-            resolve(taskStatus);
-          } else if (!subtasks.length) {
-            clearInterval(pollInterval);
-            resolve(taskStatus);
-          }
-        }
-      }, POLLING_DELAY);
-    });
-  };
-}
-
-function pollBuildAutodeployStatus(
-  accountId: number,
-  taskName: string,
-  buildId: number
-): Promise<Build> {
-  return new Promise((resolve, reject) => {
-    let maxIntervals = (30 * 1000) / POLLING_DELAY; // Num of intervals in ~30s
-
-    const pollInterval = setInterval(async () => {
-      let build: Build;
-      try {
-        const response = await getBuildStatus(accountId, taskName, buildId);
-        build = response.data;
-      } catch (e) {
-        logger.debug(e);
-        return reject(
-          new Error(
-            i18n(`${i18nKey}.pollBuildAutodeployStatusError`, { buildId })
-          )
-        );
-      }
-
-      if (!build || !build.status) {
-        return reject(
-          new Error(
-            i18n(`${i18nKey}.pollBuildAutodeployStatusError`, { buildId })
-          )
-        );
-      }
-
-      if (build.deployStatusTaskLocator || maxIntervals <= 0) {
-        clearInterval(pollInterval);
-        resolve(build);
-      } else {
-        maxIntervals -= 1;
-      }
-    }, POLLING_DELAY);
-  });
-}
-
-const pollBuildStatus = makePollTaskStatusFunc<Build>({
-  linkToHubSpot: (accountId, taskName, taskId) =>
-    uiLink(
-      `View build #${taskId} in HubSpot`,
-      getProjectBuildDetailUrl(taskName, taskId, accountId)
-    ),
-  statusFn: getBuildStatus,
-  structureFn: getBuildStructure,
-  statusText: PROJECT_BUILD_TEXT,
-  statusStrings: {
-    INITIALIZE: (name, buildId) => `Building ${chalk.bold(name)} #${buildId}`,
-    SUCCESS: (name, buildId) => `Built ${chalk.bold(name)} #${buildId}`,
-    FAIL: (name, buildId) => `Failed to build ${chalk.bold(name)} #${buildId}`,
-    SUBTASK_FAIL: (buildId, name) =>
-      `Build #${buildId} failed because there was a problem\nbuilding ${chalk.bold(
-        name
-      )}`,
-  },
-});
-
-const pollDeployStatus = makePollTaskStatusFunc<Deploy>({
-  linkToHubSpot: (accountId, taskName, taskId, deployedBuildId) =>
-    uiLink(
-      `View deploy of build #${deployedBuildId} in HubSpot`,
-      getProjectDeployDetailUrl(taskName, taskId, accountId)
-    ),
-  statusFn: getDeployStatus,
-  structureFn: getDeployStructure,
-  statusText: PROJECT_DEPLOY_TEXT,
-  statusStrings: {
-    INITIALIZE: (name, buildId) =>
-      `Deploying build #${buildId} in ${chalk.bold(name)}`,
-    SUCCESS: (name, buildId) =>
-      `Deployed build #${buildId} in ${chalk.bold(name)}`,
-    FAIL: (name, buildId) =>
-      `Failed to deploy build #${buildId} in ${chalk.bold(name)}`,
-    SUBTASK_FAIL: (deployedBuildId, name) =>
-      `Deploy for build #${deployedBuildId} failed because there was a\nproblem deploying ${chalk.bold(
-        name
-      )}`,
-  },
-});
-
-function logFeedbackMessage(buildId: number): void {
-  if (buildId > 0 && buildId % FEEDBACK_INTERVAL === 0) {
-    uiLine();
-    logger.log(i18n(`${i18nKey}.logFeedbackMessage.feedbackHeader`));
-    uiLine();
-    logger.log(i18n(`${i18nKey}.logFeedbackMessage.feedbackMessage`));
-  }
-}
-
-async function createProjectComponent(
-  component: ProjectAddComponentData,
-  name: string,
-  projectComponentsVersion: string
-): Promise<void> {
-  const i18nKey = 'commands.project.subcommands.add';
-  const componentName = name;
-
-  const configInfo = await getProjectConfig();
-
-  if (!configInfo.projectDir || !configInfo.projectConfig) {
-    logger.error(i18n(`${i18nKey}.error.locationInProject`));
-    process.exit(EXIT_CODES.ERROR);
-  }
-
-  const componentPath = path.join(
-    configInfo.projectDir,
-    configInfo.projectConfig.srcDir,
-    component.insertPath,
-    componentName
-  );
-
-  await downloadGithubRepoContents(
-    HUBSPOT_PROJECT_COMPONENTS_GITHUB_PATH,
-    component.path,
-    componentPath,
-    projectComponentsVersion
-  );
-}
-
-async function displayWarnLogs(
-  accountId: number,
-  projectName: string,
-  taskId: number,
-  isDeploy = false
-): Promise<void> {
-  let result: WarnLogsResponse | undefined;
-
-  if (isDeploy) {
-    try {
-      const { data } = await fetchDeployWarnLogs(
-        accountId,
-        projectName,
-        taskId
-      );
-      result = data;
-    } catch (e) {
-      logError(e);
-    }
-  } else {
-    try {
-      const { data } = await fetchBuildWarnLogs(accountId, projectName, taskId);
-      result = data;
-    } catch (e) {
-      logError(e);
-    }
-  }
-
-  if (result && result.logs) {
-    const logLength = result.logs.length;
-    result.logs.forEach((log, i) => {
-      logger.warn(log.message);
-      if (i < logLength - 1) {
-        logger.log('');
-      }
-    });
-  }
-}
-
-async function getProjectComponentsByVersion(
-  projectComponentsVersion: string
-): Promise<ComponentTemplate[]> {
-  const config = await fetchFileFromRepository<ProjectTemplateRepoConfig>(
-    HUBSPOT_PROJECT_COMPONENTS_GITHUB_PATH,
-    'config.json',
-    projectComponentsVersion
-  );
-
-  return config[PROJECT_COMPONENT_TYPES.COMPONENTS] || [];
-}
