+++ lib/projects/structure.ts
